Separate Batching components into own file. Temporarily include diagnostic output (which will be removed in later check-in).
