+++ vnext/ReactUWP/Base/UwpReactInstance.cpp
@@ -52,7 +52,7 @@
 #include <Modules/LocationObserverModule.h>
 #include <Modules/NativeUIManager.h>
 #include <Modules/NetworkingModule.h>
-#include <Modules/UIManagerModule.h>
+#include <Modules/BatchingUIManagerModule.h>
 #include <Modules/WebSocketModuleUwp.h>
 #include <ReactWindowsCore/IUIManager.h>
 #include <Threading/UIMessageQueueThread.h>
@@ -125,7 +125,7 @@ REACTWINDOWS_API_(std::shared_ptr<facebook::react::IUIManager>) CreateUIManager(
   viewManagers.push_back(std::make_unique<polyester::IconViewManager>(instance));
 
   // Create UIManager, passing in ViewManagers
-  return createIUIManager(std::move(viewManagers), new NativeUIManager());
+  return createBatchingUIManager(std::move(viewManagers), new NativeUIManager());
 }
 
 UwpReactInstance::UwpReactInstance(
@@ -150,7 +150,7 @@ std::vector<facebook::react::NativeModuleDescription> GetModules(
 
   modules.emplace_back(
     "UIManager",
-    [uiManager = std::move(uiManager)]() { return facebook::react::createUIManagerModule(uiManager); },
+    [uiManager = std::move(uiManager)]() { return facebook::react::createBatchingUIManagerModule(uiManager); },
     messageQueue);
 
   modules.emplace_back(
