+++ vnext/ReactWindowsCore/ReactWindowsCore.vcxproj.filters
@@ -38,6 +38,9 @@
       <Filter>Modules</Filter>
     </ClCompile>
     <ClCompile Include="LayoutAnimation.cpp" />
+    <ClCompile Include="Modules\BatchingUIManagerModule.cpp">
+      <Filter>Modules</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="CxxMessageQueue.h" />
@@ -125,6 +128,9 @@
     <ClInclude Include="..\include\ReactWindowsCore\unicode.h" />
     <ClInclude Include="LayoutAnimation.h" />
     <ClInclude Include="utilities.h" />
+    <ClInclude Include="Modules\BatchingUIManagerModule.h">
+      <Filter>Modules</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
